+++ test/e2e/test_cases/web_content/analytical_operations/sum/bubble_to_coxcomb.mjs
@@ -44,13 +44,6 @@ const testSteps = [
 
 				geometry: 'rectangle',
 				coordSystem: 'polar'
-			},
-			style: {
-				plot: {
-					marker: {
-						rectangleSpacing: '0em'
-					}
-				}
 			}
 		}),