+++ test/e2e/test_cases/ww_samples_for_presets/polar_coo_sys/46_P_R_coxcomb_nightingale_rose_chart.mjs
@@ -24,7 +24,6 @@ const testSteps = [
 			style: {
 				plot: {
 					marker: {
-						rectangleSpacing: '0',
 						fillOpacity: 0.5,
 						borderWidth: 0
 					}